sixth update of the Android studio code made by Amir Eihab El Abidou, Omar Almassri, Arianit Paso. This code is the continuation of the integration phase. We discovered that using a service class is not the best approach since that would have the app always running in the background of the users phone, thus reducing the phones battery and performance. We will instead publish the users preferred lux values in a topic that the python script will be subscribed to, and every time the user changes his preferences, the values gets saved and published. Also the switch state also gets published everytime it changes. This means that the mqttService class is now not needed and can be removed, but we will publish it so that we can have access to the logic we created inside it. Now, the mainactivity is instead subscribed to the "iot/sensor" and has implemented mqtt logic. The algorythm where the brightness level of the Lifx lamp will changed based on already existing lux value and preferred lux value will instead be implemented in the python script that will run in the Raspbarry pi, so that the dynamic change can instead be done in the raspbarry without having background activities in the phone.

+++ app/src/main/java/com/example/smartlight/TimeInterval.java
@@ -9,16 +9,15 @@ public class TimeInterval implements Serializable {
     private int startMinute;
     private int endHour;
     private int endMinute;
-    private double value;
+    private int value;
 
-    public TimeInterval(int startHour, int startMinute, int endHour, int endMinute, double value) {
+    public TimeInterval(int startHour, int startMinute, int endHour, int endMinute, int value) {
         this.startHour = startHour;
         this.startMinute = startMinute;
         this.endHour = endHour;
         this.endMinute = endMinute;
         this.value = value;
     }
-
 
     public int getStartHour() {
         return startHour;
@@ -52,23 +51,16 @@ public class TimeInterval implements Serializable {
         this.endMinute = endMinute;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(int value) {
         this.value = value;
     }
 
-
     @Override
     public String toString() {
-        return "TimeInterval{" +
-                "startHour=" + startHour +
-                ", startMinute=" + startMinute +
-                ", endHour=" + endHour +
-                ", endMinute=" + endMinute +
-                ", value=" + value +
-                '}';
+        return "TimeInterval{" + "startHour=" + startHour + ", startMinute=" + startMinute + ", endHour=" + endHour + ", endMinute=" + endMinute + ", value=" + value + '}';
     }
 }
